Hulkar

+++ src/user/user.service.ts
@@ -1,6 +1,6 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { from, map, Observable} from 'rxjs';
+import { from, Observable } from 'rxjs';
 import { Repository } from 'typeorm';
 import { UserResponse } from './dto/interface'
 import { UpdateUserDto } from './dto/update-user';
@@ -18,28 +18,39 @@ export class UserService {
     return from(this.usersRepository.find());
   }
 
-  getUser(id: number): Observable<UserResponse> {
-    return from(this.usersRepository.findOne({ where: { id } }));
+
+  async getUser(id: number): Promise<UserResponse> {
+    const user = await this.findUser(id);
+    if (!user) {
+      throw new NotFoundException('Sorry,we dont have this user.');
+    }
+    return user;
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const user = await this.findUser(id);
+    if (!user) {
+      throw new HttpException('User not found, please try again.', HttpStatus.NOT_FOUND);
+    }
     return from(this.usersRepository.delete({ id }));
   }
 
+  async findUser(id: number): Promise<User> {
+    let user: User;
+    try {
+      user = await this.usersRepository.findOneBy({ id });
+      return user || null;
+    } catch (error) {
+      return null;
+    }
+  }
   updateUser(updateUserDto: UpdateUserDto) {
-    const exist = this.userExists(updateUserDto.id);
+    const exist = this.findUser(updateUserDto.id);
     if (!exist) {
       throw new HttpException('User not found, please try again.', HttpStatus.NOT_FOUND);
     } else {
       return from(this.usersRepository.update(updateUserDto.id, updateUserDto));
     }
   }
-
-  private userExists(id: number): Observable<boolean> {
-    return from(this.usersRepository.findOne({ where: { id } })).pipe(
-      map((user: UserResponse) => {
-        return !user
-      })
-    )
   }
 }
